pyproject.toml
[tool.poetry]
name = "de-zcamp"
version = "0.1.0"
description = "Weekly lab work and capstone project submission for the DE Zoomcamp course offered by DataTalksClub."
authors = ["shamb0 <r.raajey@gmail.com>"]
readme = "README.md"
packages = [{include = "de-zcamp"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
pandas = "^2.2.0"
sqlalchemy = "^1.4.15"
python-dotenv = "^1.0.1"
numpy = "^1.26.3"
prettytable = "^3.9.0"
google-cloud-storage = "^2.14.0"
google-cloud-bigquery = "^3.17.1"

[tool.poetry.scripts]
01-run-etl-green-tripdata = "de-zcamp.10-wkly-labs.01-mod01-prerequisite.cli_01_dpipe_nyc_green_tripdata_to_localdb:cli"
02-run-etl-zone-lookup = "de-zcamp.10-wkly-labs.01-mod01-prerequisite.cli_02_dpipe_nyc_taxi_zone_lookup_to_localdb:cli"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = {version = "^24.1.1", extras = ["jupyter"]}
isort = "^5.13.2"
docformatter = "^1.7.5"
mage-ai = "^0.9.61"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '(/.git|/dist|/.venv|/venv|/build|/agent|agbenchmark/challenges)/'

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
